src/test/kotlin/dev/shtanko/algorithms/leetcode/SingleElementInSortedArrayTest.kt
/*
 * Copyright 2023 Oleksii Shtanko
 *
 * Licensed under the Apache License, Version 2.0 (the "License");
 * you may not use this file except in compliance with the License.
 * You may obtain a copy of the License at
 *
 *     http://www.apache.org/licenses/LICENSE-2.0
 *
 * Unless required by applicable law or agreed to in writing, software
 * distributed under the License is distributed on an "AS IS" BASIS,
 * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
 * See the License for the specific language governing permissions and
 * limitations under the License.
 */

package dev.shtanko.algorithms.leetcode

import java.util.stream.Stream
import org.assertj.core.api.Assertions.assertThat
import org.junit.jupiter.api.extension.ExtensionContext
import org.junit.jupiter.params.ParameterizedTest
import org.junit.jupiter.params.provider.Arguments
import org.junit.jupiter.params.provider.ArgumentsProvider
import org.junit.jupiter.params.provider.ArgumentsSource

abstract class SingleElementInSortedArrayTest<out T : SingleElementInSortedArray>(private val strategy: T) {
    private class InputArgumentsProvider : ArgumentsProvider {
        override fun provideArguments(context: ExtensionContext?): Stream<out Arguments> = Stream.of(
            Arguments.of(
                intArrayOf(1, 1, 2, 3, 3, 4, 4, 8, 8),
                2,
            ),
            Arguments.of(
                intArrayOf(3, 3, 7, 7, 10, 11, 11),
                10,
            ),
        )
    }

    @ParameterizedTest
    @ArgumentsSource(InputArgumentsProvider::class)
    fun `single non duplicate test`(nums: IntArray, expected: Int) {
        val actual = strategy.singleNonDuplicate(nums)
        assertThat(actual).isEqualTo(expected)
    }
}

class SingleElementInSortedArrayBSTest :
    SingleElementInSortedArrayTest<SingleElementInSortedArray>(SingleElementInSortedArrayBS())
